pyproject.toml
[project]
name = "video-trans"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.0.2",
    "gemini-srt-translator>=2.1.1",
    "gtts>=2.5.4",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "yt-dlp>=2025.6.9",
]
